backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
yfinance==1.2.0
gunicorn==23.0.0
whitenoise==6.8.2
python-decouple==3.8
reportlab==4.2.5
django-celery-beat==2.7.0
Pillow==11.1.0
dj-database-url==2.3.0
